Improved spawning and added in-game config
Changed up the spawning a bit and brought back the spawn everywhere.
Moved all the cfg_ objects into the Config file and also got the in-game
config to work from the mods menu.
Some of the in-game config options need a restart, so a separate
category might be needed for them.